release: 3.8.0, defaulting CompiledModel to fp32
CompiledModel's plain constructors defaulted to Precision.fp16, which produces
wrong results on every GPU measured. Strict-GPU fp32 matched a plain-CPU
reference for every model that compiled, on Adreno, Xclipse, Mali, and Apple
Metal, while fp16 matched 4 of 18, 5 of 18, 1 of 12, and 4 of 18 respectively.
Google's LiteRT Python API reproduces the Apple figures through the same
switch, so this is upstream numerical behaviour: fp16 carries about three
decimal digits of mantissa and these graphs emit pixel-space coordinates.

fromBufferWithGpuFallback already defaulted to fp32 and documented why; the
plain constructors were contradicting it. This changes numeric output and costs
roughly 15% median GPU latency, hence a minor version rather than a patch. fp16
stays available explicitly and is better chosen per model, validated on the
target GPU. No test churn: every Precision.fp16 in the suite is an explicit
argument exercising fp16 deliberately.

The README now states NPU maturity per platform rather than implying it is
uniform: macOS and iOS stable on Core ML, Android a work in progress on
Qualcomm HTP alone with no bundled runtime, Windows and Linux not implemented,
web throwing. It also records that on Android a device with no vendor runtime
is the expected case, and that NPU accuracy is model-specific, since physical
testing found fully dispatched graphs that still missed the CPU tolerance.

Adds a migration guide for tflite_flutter, which is unmaintained and which this
package forked with a source-compatible Interpreter API. The swap is two lines
plus deleting the manual native-library setup that package required.
CompiledModel is float32-only, so quantized I/O, training, custom ops, and
named signatures stay on the classic Interpreter.

+++ lib/src/compiled_model/compiled_model_selection.dart
@@ -17,10 +17,14 @@ import 'compiled_model.dart';
 ///
 /// [onGpuFallback] is invoked with the compile error when the GPU attempt fails
 /// and the model falls back to CPU. [forceCpu] skips the GPU attempt entirely.
+///
+/// [precision] defaults to [Precision.fp32]. fp16 failed CPU-reference parity
+/// for most models on every GPU architecture measured, so it is an explicit
+/// per-model opt-in rather than a default; see [CompiledModel.fromFile].
 CompiledModel compiledModelFromBufferAuto(
   Uint8List bytes, {
   Set<Accelerator> accelerators = const {Accelerator.gpu, Accelerator.cpu},
-  Precision precision = Precision.fp16,
+  Precision precision = Precision.fp32,
   TensorBufferMode tensorBufferMode = TensorBufferMode.managed,
   bool forceCpu = false,
   void Function(Object error)? onGpuFallback,
