+++ lib/src/compiled_model/compiled_model_unsupported.dart
@@ -28,7 +28,7 @@ class CompiledModel {
   static CompiledModel fromFile(
     String path, {
     Set<Accelerator> accelerators = const {Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     TensorBufferMode tensorBufferMode = TensorBufferMode.managed,
   }) {
     throw UnsupportedError('CompiledModel is not supported on this platform.');
@@ -38,7 +38,7 @@ class CompiledModel {
   static CompiledModel fromBuffer(
     Uint8List bytes, {
     Set<Accelerator> accelerators = const {Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     TensorBufferMode tensorBufferMode = TensorBufferMode.managed,
   }) {
     throw UnsupportedError('CompiledModel is not supported on this platform.');
@@ -49,7 +49,7 @@ class CompiledModel {
   static Future<CompiledModel> fromBufferAsync(
     Uint8List bytes, {
     Set<Accelerator> accelerators = const {Accelerator.cpu},
-    Precision precision = Precision.fp16,
+    Precision precision = Precision.fp32,
     TensorBufferMode tensorBufferMode = TensorBufferMode.managed,
   }) {
     throw UnsupportedError('CompiledModel is not supported on this platform.');
